requirements.txt
amqp (1.4.9)
anyjson (0.3.3)
backports.shutil-get-terminal-size (1.0.0)
billiard (3.3.0.23)
celery (3.1.23)
decorator (4.0.9)
Django (1.9.6)
django-bootstrap3 (7.0.1)
django-celery (3.1.17)
django-haystack (2.5.dev0)
elasticsearch (2.3.0)
gevent (1.1.1)
greenlet (0.4.9)
ipython (4.2.0)
ipython-genutils (0.1.0)
kombu (3.0.35)
pexpect (4.1.0)
pickleshare (0.7.2)
pip (8.1.2)
ptyprocess (0.5.1)
pytz (2016.4)
requests (2.10.0)
setuptools (21.2.1)
simplegeneric (0.8.1)
traitlets (4.2.1)
urllib3 (1.15.1)
wheel (0.29.0)
